requirements.txt
# ============================================================================
# Alchemy Trading Bot - Requirements
# ============================================================================
# Install with: pip install -r requirements.txt
#
# For minimal install (core features only):
#   pip install -r requirements.txt --only-binary :all:
#
# Note: Some packages like ta-lib require system dependencies
# ============================================================================

# ----------------------------------------------------------------------------
# WEB SERVER & API (Required for Dashboard)
# ----------------------------------------------------------------------------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
websockets>=12.0
httpx>=0.25.0

# ----------------------------------------------------------------------------
# CORE DEPENDENCIES
# ----------------------------------------------------------------------------
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ----------------------------------------------------------------------------
# TRADING & MARKET DATA
# ----------------------------------------------------------------------------
alpaca-py>=0.13.0              # New Alpaca SDK (recommended)
# alpaca-trade-api>=3.0.0      # Legacy Alpaca SDK (if needed)
yfinance>=0.2.28               # Yahoo Finance backup data

# ----------------------------------------------------------------------------
# TECHNICAL ANALYSIS
# ----------------------------------------------------------------------------
ta>=0.10.2                     # Pure Python TA (no dependencies)
pandas-ta>=0.3.14b             # Additional indicators
# ta-lib>=0.4.28               # Optional: Requires system install of TA-Lib
                               # Ubuntu: sudo apt-get install libta-lib-dev
                               # Mac: brew install ta-lib

# ----------------------------------------------------------------------------
# MACHINE LEARNING (Optional - for advanced features)
# ----------------------------------------------------------------------------
scikit-learn>=1.3.0
# xgboost>=1.7.0               # Gradient boosting
# lightgbm>=4.0.0              # Light gradient boosting
# tensorflow>=2.13.0           # Deep learning (heavy install)
# torch>=2.0.0                 # PyTorch alternative

# ----------------------------------------------------------------------------
# DATA STORAGE
# ----------------------------------------------------------------------------
# SQLite is built into Python - no extra install needed
sqlalchemy>=2.0.0              # ORM (optional)
# psycopg2-binary>=2.9.6       # PostgreSQL (if scaling up)
# redis>=4.6.0                 # Caching (if scaling up)

# ----------------------------------------------------------------------------
# UTILITIES
# ----------------------------------------------------------------------------
python-dotenv>=1.0.0           # Environment variables
python-dateutil>=2.8.2         # Date parsing
pyyaml>=6.0                    # Config files
requests>=2.31.0               # HTTP client
aiohttp>=3.8.5                 # Async HTTP

# ----------------------------------------------------------------------------
# LOGGING & MONITORING
# ----------------------------------------------------------------------------
loguru>=0.7.0                  # Better logging

# ----------------------------------------------------------------------------
# VISUALIZATION (Optional - for charts/reports)
# ----------------------------------------------------------------------------
# matplotlib>=3.7.0
# plotly>=5.15.0
# seaborn>=0.12.0

# ----------------------------------------------------------------------------
# BACKTESTING (Optional)
# ----------------------------------------------------------------------------
# backtrader>=1.9.76
# vectorbt>=0.25.0

# ----------------------------------------------------------------------------
# SENTIMENT ANALYSIS (Optional - for news features)
# ----------------------------------------------------------------------------
# newsapi-python>=0.2.7
# textblob>=0.17.1
# vaderSentiment>=3.3.2

# ----------------------------------------------------------------------------
# TESTING (Development only)
# ----------------------------------------------------------------------------
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.1.0
